Add delete post function via command /delete

+++ internal/lib/database/handlers/message/updateMessageByID.go
@@ -2,6 +2,6 @@ package message
 
 import modeldb "slm-bot-publisher/internal/lib/database/model"
 
-func (h *HandlerDBMessage) UpdateMessageByID(telegramMsgID uint64, message modeldb.Message) error {
+func (h *HandlerDBMessage) UpdateMessageByID(telegramMsgID int, message modeldb.Message) error {
 	return h.DB.Model(&modeldb.Message{}).Where("telegram_msg_id = ?", telegramMsgID).Updates(message).Error
 }
